TINY-14241: split bdd and chain tests
